# Performance improvements on ARM64